functions/snyk_report/requirements.txt
pysnyk==0.4.2
